Currently done: 1. Deleting unnecessar files and renaming files. 2. Moving the algo file to be under algorithm folder. 3. Replacing the class with the readonly static fields to constants file.
