+++ apps/server/src/algorithm/findStartActivity.ts
@@ -1,110 +1,110 @@
-import { Activity } from "../../../client/src/interfaces";
-import ICoordinate from "../../../client/src/interfaces/activity/coordinate";
-import IGroup from "../../../client/src/interfaces/activity/group";
-import { calculateDistance, randomPoint } from "../controllers/MapCalculation";
-import turf from "turf";
-import circle from "@turf/circle";
+import { Activity } from '../../../client/src/interfaces';
+import ICoordinate from '../../../client/src/interfaces/activity/coordinate';
+import IGroup from '../../../client/src/interfaces/activity/group';
+import { calculateDistance, randomPoint } from '../controllers/mapCalculation';
+import turf from 'turf';
+import circle from '@turf/circle';
 
 export const findStartAlgoPoint = (
-  cityCenter: ICoordinate,
-  radius: number,
-  selectedActivities: Activity[]
+    cityCenter: ICoordinate,
+    radius: number,
+    selectedActivities: Activity[]
 ): ICoordinate => {
-  if (selectedActivities.length != 0) {
-    return findBestCoverPoint(radius, selectedActivities);
-  } else {
-    return findFarthestPoint(cityCenter, radius);
-  }
+    if (selectedActivities.length != 0) {
+        return findBestCoverPoint(radius, selectedActivities);
+    } else {
+        return findFarthestPoint(cityCenter, radius);
+    }
 };
 
 const findBestCoverPoint = (
-  radius: number,
-  selectedActivities: Activity[]
+    radius: number,
+    selectedActivities: Activity[]
 ): ICoordinate => {
-  // Divide all selected activities to groups
-  let activitiesGroups: IGroup[] = groupPointsByRadius(
-    selectedActivities,
-    radius
-  );
+    // Divide all selected activities to groups
+    let activitiesGroups: IGroup[] = groupPointsByRadius(
+        selectedActivities,
+        radius
+    );
 
-  // Find the group that has maximum selected activities
-  let maxActivities = 0;
-  let maxGroupIndex = -1;
-  activitiesGroups.map((activity, index) => {
-    if (activity.selectedActivities.length > maxActivities) {
-      maxActivities = activity.selectedActivities.length;
-      maxGroupIndex = index;
-    }
-  });
+    // Find the group that has maximum selected activities
+    let maxActivities = 0;
+    let maxGroupIndex = -1;
+    activitiesGroups.map((activity, index) => {
+        if (activity.selectedActivities.length > maxActivities) {
+            maxActivities = activity.selectedActivities.length;
+            maxGroupIndex = index;
+        }
+    });
 
-  // For this group, find the center point
-  return findCenterPoint(activitiesGroups[maxGroupIndex].selectedActivities);
+    // For this group, find the center point
+    return findCenterPoint(activitiesGroups[maxGroupIndex].selectedActivities);
 };
 
 const groupPointsByRadius = (
-  selectedActivities: Activity[],
-  radius: number
+    selectedActivities: Activity[],
+    radius: number
 ): IGroup[] => {
-  let groups: IGroup[] = [];
+    let groups: IGroup[] = [];
 
-  for (let currentSelectedActivity of selectedActivities) {
-    let closestGroup: IGroup | null = null;
-    let closestDistance = Infinity;
+    for (let currentSelectedActivity of selectedActivities) {
+        let closestGroup: IGroup | null = null;
+        let closestDistance = Infinity;
 
-    for (let group of groups) {
-      const distance = calculateDistance(
-        group.centerActivity.position,
-        currentSelectedActivity.position
-      );
+        for (let group of groups) {
+            const distance = calculateDistance(
+                group.centerActivity.position,
+                currentSelectedActivity.position
+            );
 
-      if (distance <= radius && distance < closestDistance) {
-        closestGroup = group;
-        closestDistance = distance;
-      }
+            if (distance <= radius && distance < closestDistance) {
+                closestGroup = group;
+                closestDistance = distance;
+            }
+        }
+
+        if (closestGroup) {
+            closestGroup.selectedActivities.push(currentSelectedActivity);
+        } else {
+            groups.push({
+                centerActivity: currentSelectedActivity,
+                selectedActivities: [currentSelectedActivity],
+            });
+        }
     }
 
-    if (closestGroup) {
-      closestGroup.selectedActivities.push(currentSelectedActivity);
-    } else {
-      groups.push({
-        centerActivity: currentSelectedActivity,
-        selectedActivities: [currentSelectedActivity],
-      });
-    }
-  }
-
-  return groups;
+    return groups;
 };
 
 const findCenterPoint = (activities: Activity[]): ICoordinate => {
-  const location: number[][][] = [[]];
+    const location: number[][][] = [[]];
 
-  activities.map((activity) => {
-    location[0].push([activity.position.lng, activity.position.lat]);
-  });
-  location[0].push([activities[0].position.lng, activities[0].position.lat]);
+    activities.map(activity => {
+        location[0].push([activity.position.lng, activity.position.lat]);
+    });
+    location[0].push([activities[0].position.lng, activities[0].position.lat]);
 
-  const polygon = turf.polygon(location);
+    const polygon = turf.polygon(location);
 
-  return randomPoint(turf.bbox(polygon));
+    return randomPoint(turf.bbox(polygon));
 };
 
 const findFarthestPoint = (
-  cityCenter: ICoordinate,
-  radius: number
+    cityCenter: ICoordinate,
+    radius: number
 ): ICoordinate => {
-  let circle1 = circle(
-    {
-      type: "Feature",
-      geometry: {
-        type: "Point",
-        // Note order: longitude, latitude.
-        coordinates: [cityCenter.lng, cityCenter.lat],
-      },
-      properties: {},
-    },
-    radius
-  );
+    let circle1 = circle(
+        {
+            type: 'Feature',
+            geometry: {
+                type: 'Point',
+                // Note order: longitude, latitude.
+                coordinates: [cityCenter.lng, cityCenter.lat],
+            },
+            properties: {},
+        },
+        radius
+    );
 
-  return randomPoint(turf.bbox(circle1));
+    return randomPoint(turf.bbox(circle1));
 };
