--- apps/server/src/bestDayRouteAlgo/bestDayRouteConstants.ts
@@ -1,7 +1,0 @@
-export class BestDayRouteConsts {
-    static readonly EarthRadius: number = 6371;
-    static readonly SplitToQuarter: number = 4;
-    static readonly DefaultDistance: number = 1.0;
-    static readonly AverageSpeed: number = 27.5;
-    static readonly MinimalTravelTimeSlot: number = 1;
-}
