+++ apps/server/src/constants/algorithm.ts
@@ -1,0 +1,3 @@
+export const SplitToQuarter: number = 4;
+export const AverageSpeed: number = 27.5;
+export const MinimalTravelTimeSlot: number = 1;
