+++ apps/server/src/controllers/mapPlace.ts
@@ -1,6 +1,10 @@
 import { schema, TABLES } from '../../../../utils';
 
-function getPlaceSQLQuery(countryCode: string, stateName?: string, cityName?: string): string {
+function getPlaceSQLQuery(
+    countryCode: string,
+    stateName?: string,
+    cityName?: string
+): string {
     let sqlQuery = `(SELECT id FROM ${schema}.${TABLES.ADDRESS} WHERE country_code='${countryCode}'`;
 
     if (stateName !== undefined) {
